✨ 기능 추가 : 헬스장 별 운동기구 수정 삭제 기능, 테스트 추가

+++ healthtart/src/main/java/com/dev5ops/healthtart/equipment_per_gym/repository/EquipmentPerGymRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EquipmentPerGymRepository extends JpaRepository<EquipmentPerGym, Integer> {
+public interface EquipmentPerGymRepository extends JpaRepository<EquipmentPerGym, Long> {
 }
